moved Result class to gitter_api.dart file
