+++ .todo
@@ -1,0 +1,5 @@
+- Improve Message class
+    - [] Add "threadMessageCount"
+    - [] Add isParent getter
+    - [] Add isChild getter
+    - [] Add isEdited getter
