+++ lib/models.dart
@@ -4,6 +4,5 @@ export 'models/mention.dart';
 export 'models/message.dart';
 export 'models/permission.dart';
 export 'models/provider.dart';
-export 'models/result.dart';
 export 'models/room.dart';
 export 'models/user.dart';
